Replaced explicit specializations of the `binder` struct template
Modern C++ allows for expressing this kind of construct in a more compact and flexible way using variadic templates and `std::enable_if` based _SFINAE_.
This particular implementation traverses the columns of the `database_binder` recursively and accumulates the gathered values in a variadic parameter pack. The parameter pack is then expanded into `std::move` instantiations during the final recursion stage.

The primary advantage of this approach to implementing `binder` lies in the elimination of manual code repetition and as such the reduction of e.g. _C&P_ introduced bugs. Furthermore it better fulfills this libraries goal of being a _modern C++ Sqlite library_.

This implementation was tested using both _gcc 4.9.2 20150304_ and _Clang 3.6.0_ on Linux.