Merged extraction operators into member function templates
The `operator>>` overloads of `sqlite::database_binder` can be merged into two member function templates where one handles value extraction and the other handles passing all extracted values to another function.
This required the implementation of a `is_sqlite_value` _template to distinguish between the overloads. Note that I would have much rather implemented this as a check for function arguments but did not as of now come up with a satisfying way to do that.

During the course of merging the overloads a new `enable_if` template alias was added to simplify _SFINAE_ constructs.